effects: isNotFound is node's after all, and say what moved
Review found `effects/todo/node-module-layering.md`, an open design that
had already decided two of these questions. It was right about one and
outdated about the other, so this reconciles both rather than leaving
contradictory guidance for whoever implements it.

`isNotFound` moves back to `effects/node`, with its proofs. It reads
`ENOENT`, a POSIX filesystem code a host without a filesystem never
reports, so it is a node predicate — "none of it is node's" was
overreach for that one function. Being about a host failure does not
make a thing host-agnostic; being about no host in particular does.

The issue's other ruling — that `IoResult` must not move to the core —
is overturned in the file with the reason: it was reasoned without a
second host, and a browser interpreter cannot declare `fetch` or
`import` without that alias existing somewhere shared. Its stated test
("not an effect constructor or combinator") also did not describe the
file it was applied to, where `NotImplemented` already lives. The
one-site `fjs/media/type` cleanup it proposed survives and stays open.

Also records why a re-export here is not the "no shims behind" case
that issue rules out: node's own operations are declared in these
types, so it re-exports what it genuinely uses.

Changelog entry added — the core module gains public exports even
though no old import path changed.
