+++ changelog/unreleased/1753.md
@@ -1,0 +1,4 @@
+- `effects`: the vocabulary every operation is declared in — `OpResult`,
+  `IoChannel`, `IoError`, `IoErrorInfo`, `IoResult`, and the `ioError` /
+  `toIoError` constructors — is now importable from the core module, not only
+  through `effects/node`, which re-exports it unchanged
