+++ fjs/effects/node/module.f.mjs
@@ -23,21 +23,26 @@ import { codePointListToString } from '../../text/utf16/module.f.mjs'
 import { reverse } from '../../types/list/module.f.mjs'
 import { length } from '../../types/bit_vec/module.f.mjs'
 import { error as resultError, ok as resultOk, unwrap } from '../../types/result/module.f.mjs'
-import { do_, ioError, isNotFound, pure, toIoError } from '../module.f.mjs'
+import { do_, ioError, pure, toIoError } from '../module.f.mjs'
 import {
     mapStep as ioMapStep, pureError, pureOk, resultMapStep, resultStep, step as ioStep,
 } from '../module.f.mjs'
 
 /**
- * The host-error vocabulary — `ioError`, `toIoError`, `isNotFound` — is
- * declared in [`../module.f.mjs`](../module.f.mjs) beside the effect
- * representation, because none of it is node's: normalizing a thrown value and
- * telling "the runner cannot" from "the host tried and failed" are what any
- * host's interpreter does. Re-exported here so the modules that reach for them
- * through the node module keep working, and so an operation's declaration and
- * its failure constructor still read as one vocabulary.
+ * `ioError` and `toIoError` are declared in
+ * [`../module.f.mjs`](../module.f.mjs) beside the effect representation,
+ * because neither is node's: normalizing a thrown value into serializable
+ * effect data is what any host's interpreter does at its `catch`. They are
+ * re-exported here so the modules that reach for them through the node module
+ * keep working, and so an operation's declaration and its failure constructor
+ * still read as one vocabulary.
+ *
+ * {@link isNotFound} stayed, and the difference is the test for where any of
+ * this belongs: it reads `ENOENT`, a POSIX filesystem code that no browser
+ * ever reports. Being about a *host failure* does not make a thing
+ * host-agnostic — being about no host in particular does.
  */
-export { ioError, isNotFound, toIoError }
+export { ioError, toIoError }
 
 /**
  * The host a {@link Listen} refuses.
@@ -84,6 +89,29 @@ export const emptyHostError = ioError({
     code: emptyHostCode,
     message: emptyHostMessage,
 })
+
+/**
+ * True if `e` is a "file or directory does not exist" (`ENOENT`) error.
+ *
+ * Node's filesystem rejections are `Error`s carrying `code: 'ENOENT'`, which
+ * {@link toIoError} keeps; the virtual interpreter reports the same code for
+ * absent paths. Lets callers swallow only the missing-path case (e.g. a fresh
+ * store) while propagating genuine failures (permissions, corruption) rather
+ * than masking them.
+ *
+ * A {@link NotImplemented} is never "not found": a runner that cannot perform
+ * the operation has not looked for the path at all, so the two must not
+ * collapse into one benign branch — which is exactly what a bare `unknown`
+ * error channel used to allow.
+ *
+ * **It belongs to this layer, unlike the constructors above.** `ENOENT` is a
+ * POSIX filesystem code; a host without a filesystem never reports one, so a
+ * shared `isNotFound` would be a node predicate wearing a host-agnostic name.
+ *
+ * @type {(e: IoChannel) => boolean}
+ */
+export const isNotFound = ([tag, payload]) =>
+    tag === 'ioError' && payload.code === 'ENOENT'
 
 /**
  * `NodeOp`'s commands as data, so a runner that implements only part of them
